chore: use startText
Start card text now come from server and are no longer stored on the
frontend app.
startText is formatted as markdown and react-markdown renders it with the given
components.
