+++ src/locales/fa/messages.po
@@ -55,10 +55,6 @@ msgstr "خطایی رخ داده. دوباره تلاش کنید."
 msgid "Are you sure you want to delete this project review?"
 msgstr "آیا از حذف  این پروژه مطمین هستید؟"
 
-#: src/pages/start-page/StartReviewCard.tsx:27
-msgid "Begin my review"
-msgstr "شروع ارزیابی من"
-
 #: src/shared/sticky-action-bar/StickyActionBar.stories.tsx:11
 msgid "Cancel"
 msgstr "لغو"
@@ -83,8 +79,7 @@ msgid "Dashboard"
 msgstr "پیشخوان"
 
 #: src/pages/start-page/NoPeerReviewPage.tsx:31
-#: src/pages/start-page/ResultStartCard.tsx:18
-#: src/pages/start-page/StartReviewCard.tsx:19
+#: src/pages/start-page/StartReviewCard.tsx:26
 msgid "Dear {name}, Hello"
 msgstr "{name} عزیز، سلام"
 
@@ -358,10 +353,6 @@ msgstr "مشاهده کمتر"
 msgid "See More"
 msgstr "مشاهده بیشتر"
 
-#: src/pages/start-page/ResultStartCard.tsx:25
-msgid "See Review Result"
-msgstr "مشاهده نتایج ارزیابی"
-
 #: src/shared/rating/Rating.stories.tsx:13
 #: src/shared/rating/Rating.stories.tsx:23
 #: src/shared/rating/Rating.stories.tsx:33
@@ -395,6 +386,10 @@ msgstr "خطایی رخ داد."
 #: src/shared/error/OverlayError.tsx:9
 msgid "Something went wrong. Please try again"
 msgstr "خطایی رخ داد. لطفا مجددا تلاش نمایید"
+
+#: src/pages/start-page/StartReviewCard.tsx:33
+msgid "Start"
+msgstr "شروع"
 
 #: src/shared/full-page-error/FullPageError.stories.tsx:15
 msgid "Strengths and Weaknesses"
