--- src/pages/start-page/PeerStartReviewContent.mdx
@@ -1,13 +1,0 @@
-لطفا این نکات رو در نظر داشته باش:
-
-<br />
-
-۱) باتوجه به اهمیت بازخوردها برای همکارات، لطفا وقت و حوصله‌ی کافی رو در این ارزیابی اختصاص بده.
-
-۲) باتوجه به امکان ویرایش ارزیابی، می‌تونی طی چند روز، مشاهداتت رو کامل و غنی‌‌ کنی تا همکارت ارزیابی دقیق‌تری داشته باشه.
-
-۳) برای نوشتن “مشاهده”، پیشنهاد می‌کنم اول، راهنمای ارتباط بدون خشونت یا NVC موجود در سامانه رو مطالعه کنی و برای بازخورددادن به همکارت ازش استفاده کن.
-
-<br />
-
-با سپاس
