--- src/pages/start-page/ResultStartPageContent.mdx
@@ -1,2 +1,0 @@
-با کمک شما تا الان دو فاز از ارزیابی عملکرد انجام شده. توی فاز اول خودت در مورد شایستگی‌های عملکردی، ویژگی‌های بارز و دستاوردها، نظراتت رو بهمون انتقال دادی. و در فاز دوم، همکارانی که در دستاوردها بهمون معرفی کردی در همین سه موضوع در مورد تو به ما نظرشون رو گفتن.  
-الان وقتشه که مرور کنی و ببینی نظر خود و دیگران در موردت چی بوده، پس با ما همراه باش.
