--- src/pages/start-page/SelfStartReviewContent.mdx
@@ -1,11 +1,0 @@
-این بخش از مرور عملکرد که در ادامه می‌بینید فقط **Self-Review** هست و سه مرحله داره که برای نوشتن هر کدوم به تمرکز نیاز دارید.
-
-
->**مرحله‌ی ۱:** خودتون رو در ۶ شایستگی اساسی ارزیابی کنید. (زمان پیش‌بینی شده: ۲۰ دقیقه)
->
->**مرحله‌ی ۲:** نقاط قوت و ضعفتون رو ذکر کنید. (زمان پیش‌بینی شده: ۱۰ دقیقه)
->
->**مرحله‌ی ۳:** به ازای هر پروژه دستاوردهاتون رو بنویسید و افرادی که قراره شما رو مرور کنند، مشخص کنید. (زمان پیش‌بینی شده به ازای هر پروژه: ۱۵ دقیقه)
-
-
-تکمیل کردن این پرسشنامه‌ها نهایت ۱ ساعت از شما وقت می‌گیره. ممنون می‌شیم وقت بذاری و اونا رو تکمیل کنی. ضمنا می‌تونی هر وقت خواستی بخشی از فرم‌ها رو ذخیره کنی و بعدا دوباره به تکمیل کردنشون بپردازی.
