+++ src/pages/start-page/StartReviewCard.tsx
@@ -1,27 +1,25 @@
-import React, { useCallback, useContext } from 'react';
+import ReactMarkdown from 'react-markdown';
+import React, { useCallback } from 'react';
 import { Box, Button, Card, CardContent, CardHeader, Grid, Theme, makeStyles } from '@material-ui/core';
 import { CSSProperties } from '@material-ui/core/styles/withStyles';
 import { FCProps } from 'src/shared/types/FCProps';
-import { MDXComponentProps, MDXContext } from '@mdx-js/react';
 import { Styles } from 'src/shared/types/Styles';
 import { getUserLabel } from 'src/shared/utils/getUserLabel';
 import { i18n } from '@lingui/core';
+import { useAppSettings } from 'src/core/settings';
 import { useAuthGuardUser } from 'src/core/auth';
 
-import { BlackQuote } from './BlackQuote';
+import { renderers } from './renderers';
 import { useStartReviewMutation } from './start-review.mutation';
 
-interface OwnProps {
-  Content: React.ComponentType<MDXComponentProps>;
-}
+interface OwnProps {}
 
 type Props = FCProps<OwnProps> & StyleProps;
 
 export function StartReviewCard(props: Props) {
-  const { Content } = props;
   const classes = useStyles(props);
-  const components = useContext(MDXContext);
   const user = useAuthGuardUser();
+  const { startText } = useAppSettings();
   const startReviewMutation = useStartReviewMutation();
 
   const startReview = useCallback(() => {
@@ -32,13 +30,12 @@ export function StartReviewCard(props: Props) {
     <Card classes={{ root: classes.root }}>
       <CardHeader title={i18n._('Dear {name}, Hello', { name: getUserLabel(user) })} />
       <CardContent>
-        <Content components={{ ...components, blockquote: BlackQuote }} />
-
+        <ReactMarkdown renderers={renderers}>{startText ?? ''}</ReactMarkdown>
         <Grid container justify="center">
           <Grid item>
             <Box marginTop={3}>
               <Button variant="contained" color="secondary" size="large" onClick={startReview}>
-                {i18n._('Begin my review')}
+                {i18n._('Start')}
               </Button>
             </Box>
           </Grid>
