+++ src/pages/start-page/StartReviewPage.tsx
@@ -1,21 +1,18 @@
 import React from 'react';
 import { Box, Container } from '@material-ui/core';
 import { FCProps } from 'src/shared/types/FCProps';
-import { importMDX } from 'mdx.macro';
 
-import { ResultStartCard } from './ResultStartCard';
-
-const Content = importMDX.sync('./ResultStartPageContent.mdx');
+import { StartReviewCard } from './StartReviewCard';
 
 interface OwnProps {}
 
 type Props = FCProps<OwnProps>;
 
-export default function ResultStartPage(props: Props) {
+export default function StartReviewPage(props: Props) {
   return (
     <Container maxWidth="md">
       <Box marginTop={15}>
-        <ResultStartCard Content={Content} />
+        <StartReviewCard />
       </Box>
     </Container>
   );
