+++ src/pages/start-page/renderers.tsx
@@ -1,0 +1,24 @@
+import React from 'react';
+import type { ReactMarkdownPropsBase } from 'react-markdown';
+import { Typography, lighten, styled } from '@material-ui/core';
+
+const StyledBlackQuote = styled('blockquote')(({ theme }) => ({
+  backgroundColor: lighten(theme.palette.primary.main, 0.85),
+  borderRadius: theme.spacing(0.5),
+  padding: theme.spacing(4),
+  marginRight: 0,
+  marginLeft: 0,
+  marginTop: theme.spacing(3),
+  marginBottom: theme.spacing(3),
+  '& p': {
+    lineHeight: '2.5em !important',
+  },
+}));
+
+export const renderers: ReactMarkdownPropsBase['renderers'] = {
+  paragraph: Typography,
+  blockquote: StyledBlackQuote,
+  heading: ({ children, level }: { children: React.ReactNode; level: 1 | 2 | 3 | 4 | 5 | 6 }) => (
+    <Typography variant={('h' + level) as 'h1' | 'h2' | 'h3' | 'h4' | 'h5' | 'h6'}>{children}</Typography>
+  ),
+};
